+++ api/src/core/interfaces/helpers/hash-function.ts
