--- api/src/core/usecases/CreateAdmin/index.ts
@@ -1,2 +1,0 @@
-export { CreateAdmin, ICreateAdminProps } from './CreateAdmin'
-export { ICreateAdminDTOInput, ICreateAdminDTOOutput } from './CreateAdminDTO'
