+++ api/src/core/usecases/create-admin/index.ts
@@ -1,0 +1,2 @@
+export { CreateAdmin, ICreateAdminProps } from './create-admin'
+export { ICreateAdminDTOInput, ICreateAdminDTOOutput } from './create-admin-dto'
